feat(fb): add Bochs VBE framebuffer driver and /dev/fb server
Implement the K1 framebuffer backend: a BochsArch FbArch backend in
drivers and an fb server serving /dev/fb over the CDEV protocol.

- drivers: BochsArch probes PCI 1234:1111 via the SYS_DEVIO hook,
  maps the framebuffer + register BARs through VM_MAP_PHYS, mode-sets
  to 1024x768x32, and fills a red/green/blue test pattern. The VBE
  register file lives at MMIO offset 0x500 of the register BAR (QEMU
  PCI_VGA_BOCHS_OFFSET; offset 0 holds the EDID), not at the BAR base
  -- without the offset the mode-set writebacks read 0xFFFF and init
  failed silently. BochsArch is #[repr(C)] so the field layout is
  deterministic for gdb reads and the ioctl ABI.
- servers: fb server (boot proc 16) serving CDEV open/close/read/write
  and grant-based ioctls, wired via map_driver(b"fb", 19, FB_PROC_NR)
  and RS boot_svcs. do_map_phys reads the phys as u32 so BARs above
  2 GiB don't sign-extend.
- boot: /sbin/fb added to the boot image, manifest, and /dev/fb
  (major 19 minor 0); kernel-boot/table/arch-common carry the new
  proc and device numbers.
- tests: 550 drivers tests (fake PCI config + physmap covering the
  port and MMIO paths); verified on QEMU via screendump (1024x768
  red/green/blue) and cat /dev/fb. Probe scripts in tools/.

+++ crates/servers/src/vm/mod.rs
@@ -1369,7 +1369,10 @@ fn do_remap(msg: &mut Message) -> i32 {
 fn do_map_phys(msg: &mut Message) -> i32 {
     let target = unsafe { msg.m_payload.m1.m1i1 };
     let len = unsafe { msg.m_payload.m1.m1i2 };
-    let phys = unsafe { msg.m_payload.m1.m1i3 } as u64;
+    // The physical address travels as 32-bit bits (device BARs live below
+    // 4 GiB); reading the i32 field directly would sign-extend addresses
+    // above 2 GiB into the 64-bit range.
+    let phys = (unsafe { msg.m_payload.m1.m1i3 } as u32) as u64;
 
     if len <= 0 {
         return EINVAL;
